Add tailwind setup and new animated hero
