+++ frontend/package.json
@@ -13,10 +13,14 @@
     "chart.js": "^4.4.9",
     "vue": "^3.5.13",
     "vue-chartjs": "^5.3.2",
-    "vue-router": "^4.5.1"
+    "vue-router": "^4.5.1",
+    "@vueuse/motion": "^3.0.3"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.2.2",
-    "vite": "^6.3.1"
+    "vite": "^6.3.1",
+    "tailwindcss": "^3.4.4",
+    "autoprefixer": "^10.4.17",
+    "postcss": "^8.4.35"
   }
 }
