+++ frontend/src/main.js
@@ -1,4 +1,6 @@
 import { createApp } from 'vue'
 import App from './App.vue'
+import './style.css'
+import { MotionPlugin } from '@vueuse/motion'
 
-createApp(App).mount('#app')
+createApp(App).use(MotionPlugin).mount('#app')
